Moved non-page components to src/components
